[Summer of code] Stand alone a new controller module (#4333)
* feature: add new module controller

* feature: add heartbeat manager

* feature: link requestProcessor and heartbeatmanager

* feature: add controllerManager and startup

* feature: remove namesrv's duplicate controller code

* review code
